requirements.txt
# Email processing
imaplib2>=2.57.0
email-validator>=1.1.3

# LLM client
requests>=2.28.0
openai>=0.27.0  # For OpenAI-compatible API format

# Utilities
python-dotenv>=0.20.0
schedule>=1.1.0
pyyaml>=6.0
tqdm>=4.64.0
python-crontab>=2.6.0  # For cron job setup
